requirements/requirements-dev.txt
ipykernel==6.29.5
python-dotenv==1.0.1
pynacl==1.5.0
loguru==0.7.2
boto3==1.35.63